Refactoring similar components into single one
RecordMatchingSystemSelector and RecordSetSelector have now been consolidated
into CollapsibleSelector. Additionally, CollapsibleSelector is now a
presentational component. It is not a redux container. I'm trying to reduce the
number of containers in the app.
